requirements.txt
Flask
Flask-SocketIO
Flask-Cors
pymongo
Flask-PyMongo
pyserial
pyFirmata
google-generativeai
python-dotenv